delete account

+++ grails-app/controllers/com/bookkeeping/UserController.groovy
@@ -1,18 +1,25 @@
 package com.bookkeeping
 
 import com.bookkeeping.dao.UserDaoService
+import com.bookkeeping.dao.UserRoleDaoService
 import com.bookkeeping.security.User
+import com.bookkeeping.security.UserRole
+import grails.gorm.transactions.Transactional
+import grails.plugin.springsecurity.SpringSecurityService
 import grails.plugin.springsecurity.annotation.Secured
 import grails.validation.ValidationException
 import static org.springframework.http.HttpStatus.*
 
+@Secured('ROLE_LIBRARIAN')
 class UserController {
 
     UserDaoService userDaoService
+    UserRoleDaoService userRoleDaoService
+
+    SpringSecurityService springSecurityService
 
     static allowedMethods = [save: "POST", update: "PUT", delete: "DELETE"]
 
-    @Secured('ROLE_LIBRARIAN')
     def index(Integer max) {
         params.max = Math.min(max ?: 10, 100)
 
@@ -21,20 +28,16 @@ class UserController {
          count: userList.getTotalCount()]
     }
 
-    @Secured('ROLE_LIBRARIAN')
     def show(Long id) {
         respond userDaoService.get(id)
     }
 
-    @Secured('ROLE_LIBRARIAN')
     def create() {
         respond new User(params)
     }
 
-    @Secured('ROLE_LIBRARIAN')
     def save(User user) {
 
-        println "save ${user}"
         if (user == null) {
             notFound()
             return
@@ -45,11 +48,9 @@ class UserController {
             return
         }
 
-        println "here"
-
         try {
             userDaoService.save(user)
-            userDaoService.setRoles(user, params.getList('roles'))
+            userRoleDaoService.setRoles(user, params.getList('roles'))
         } catch (ValidationException e) {
             respond user.errors, view:'create'
             return
@@ -58,12 +59,10 @@ class UserController {
         redirect action:"index", params: [sort: "dateCreated", order: "desc"]
     }
 
-    @Secured('ROLE_LIBRARIAN')
     def edit(Long id) {
         respond userDaoService.get(id)
     }
 
-    @Secured(['ROLE_LIBRARIAN'])
     def update(User user) {
 
         if (user == null) {
@@ -78,7 +77,7 @@ class UserController {
 
         try {
             userDaoService.save(user)
-            userDaoService.setRoles(user, params.getList('roles'))
+            userRoleDaoService.setRoles(user, params.getList('roles'))
         } catch (ValidationException e) {
             respond user.errors, view:'edit'
             return
@@ -87,13 +86,22 @@ class UserController {
         redirect action:"index", params: [sort: "lastUpdated", order: "desc"]
     }
 
-    @Secured('ROLE_LIBRARIAN')
+    @Transactional
     def delete(Long id) {
+
+        User secUser = springSecurityService.currentUser as User
+
         if (id == null) {
             notFound()
             return
         }
 
+        if (id == secUser.id) {
+            badRequest()
+            return
+        }
+
+        userRoleDaoService.removeRoles(userDaoService.get(id))
         userDaoService.delete(id)
 
         request.withFormat {
@@ -105,35 +113,6 @@ class UserController {
         }
     }
 
-    @Secured(['IS_AUTHENTICATED_ANONYMOUSLY'])
-    def register() {
-        respond new User(params)
-    }
-
-    @Secured(['IS_AUTHENTICATED_ANONYMOUSLY'])
-    def saveUser(User user) {
-
-        if(!user.validate()) {
-            render(view: 'register', model: [user: user])
-            return
-        }
-
-        try {
-            userDaoService.saveUser(user)
-        } catch (ValidationException e) {
-            respond user.errors, view:'register'
-            return
-        }
-
-        request.withFormat {
-            form multipartForm {
-                flash.message = message(code: 'default.created.message', args: [message(code: 'user.label', default: 'User')])
-                redirect uri: '/login/auth'
-            }
-            '*' { respond user, [status: CREATED] }
-        }
-    }
-
     protected void notFound() {
         request.withFormat {
             form multipartForm {
@@ -141,6 +120,12 @@ class UserController {
                 redirect action: "index", method: "GET"
             }
             '*'{ render status: NOT_FOUND }
+        }
+    }
+
+    protected void badRequest() {
+        request.withFormat {
+            '*'{ render status: BAD_REQUEST }
         }
     }
 }
